requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
asgiref==3.8.1
sqlparse==0.5.2
gunicorn==23.0.0
drf-yasg==1.21.10
djoser
djangorestframework_simplejwt==5.5.0
django-filter